// src/core/interfaces/dashboard.ts
export interface Dashboard {
  id: string;
  name: string;
  widgets: Widget[];
  v?: string;
}

export interface SeriesConfig {
  metric: string;
  host: string;
  metrics?: string[];
  hosts?: string[];
  yAxis?: 'left' | 'right';
  chartType: 'line' | 'bar' | 'area' | 'pie';
  aggregation: 'none' | 'sum' | 'avg';
  stacked: boolean;
}

export interface Widget {
  id: string;
  type: 'kpi' | 'chart';
  title: string;
  metrics: string[];
  hosts: string[];
  x: number;
  y: number;
  w: number;
  h: number;
  chartType?: 'line' | 'bar' | 'area' | 'pie' | 'mixed';
  seriesConfig?: Record<string, SeriesConfig>;
  aggregation?: 'none' | 'sum' | 'avg';
  stacked?: boolean;
  forceNewline?: boolean;
}